requirements.txt
asgiref==3.3.1
Django==3.1.5
django-environ==0.4.5
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
Pillow==8.1.0
python-resize-image==1.1.19
Fiona==1.8.18
pyproj==3.0.0.post1
Shapely==1.7.1
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
django-simple-captcha==0.5.13
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
djangorestframework-gis==0.17
